Added dynamic models load to `documents_registry`
